Split DRF backend and integration tests
